+++ packages/node-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @c15t/node-sdk
 
+## 1.5.0-canary-20250722133547
+
+### Patch Changes
+
+- Updated dependencies [35b5fc3]
+  - @c15t/backend@1.5.0-canary-20250722133547
+
+## 1.5.0-canary-20250722085128
+
+### Patch Changes
+
+- Updated dependencies [74e1825]
+  - @c15t/backend@1.5.0-canary-20250722085128
+
 ## 1.4.2
 
 ### Patch Changes