+++ packages/react/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @c15t/react
 
+## 1.5.0-canary-20250722151007
+
+### Patch Changes
+
+- 5decee3: chore(react): remove un-used dependency
+
+## 1.5.0-canary-20250722133547
+
+### Patch Changes
+
+- 35b5fc3: fix: removed package.json imports for Next 13 compatibility
+- Updated dependencies [35b5fc3]
+  - c15t@1.5.0-canary-20250722133547
+  - @c15t/node-sdk@1.5.0-canary-20250722133547
+
+## 1.5.0-canary-20250722085128
+
+### Patch Changes
+
+- Updated dependencies [35f052c]
+  - c15t@1.5.0-canary-20250722085128
+  - @c15t/node-sdk@1.5.0-canary-20250722085128
+
 ## 1.4.4
 
 ### Patch Changes